Wrongly delete nfcp_bess_generator; recover it and rename as lemur_bess_generator
